Created signInLocal,signUpLocal,signInJwt functions
signInLocal takes signInDto and checks if the credentials match the database. If yes, it issues a jwt token (signInJwt()) and returns that.
signUpLocal takes signUpDto and checks if the credentials already exists in the database. If not, it will create a new user ard issue a jwt token and return it
signInJwt takes userId and email and issues a jwt token and returns it